+++ frontend/src/app/app-routing.module.ts
@@ -1,14 +1,11 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AppComponent } from './app.component';
 import { AdminPageComponent } from './layouts/admin/admin-page/admin-page.component';
-import {HomeComponent} from './layouts/home/home.component';
-import {ProjectComponent} from './project/project.component';
+import {ProjectComponent} from './layouts/project/project.component';
 
 const routes: Routes = [
-  // { path: '', component:  },
 	{ path: '', 	component: ProjectComponent },
-  { path: 'admin', component: AdminPageComponent}
+	{ path: 'admin', component: AdminPageComponent}
 ];
 
 @NgModule({
