+++ frontend/src/app/app.module.ts
@@ -11,11 +11,12 @@ import { AdminPageComponent } from './layouts/admin/admin-page/admin-page.compon
 import { AdminFooterComponent } from './layouts/admin/admin-page/footer/footer.component';
 import { AdminHeaderComponent} from './layouts/admin/admin-page/header/header.component';
 import { AdminSidebarComponent} from './layouts/admin/admin-page/sidebar/sidebar.component';
-import { ProjectComponent } from './project/project.component';
-import { FooterComponent } from './project/footer/footer.component';
-import {MainComponent} from './project/main/main.component';
-import {SidebarComponent} from './project/sidebar/sidebar.component';
-import {HeaderComponent} from './project/header/header.component';
+import { ProjectComponent } from './layouts/project/project.component';
+import { FooterComponent } from './layouts/project/footer/footer.component';
+import {MainComponent} from './layouts/project/main/main.component';
+import {SidebarComponent} from './layouts/project/sidebar/sidebar.component';
+import {HeaderComponent} from './layouts/project/header/header.component';
+import { AdminContentComponent } from './layouts/admin/admin-page/content/content.component';
 
 @NgModule({
 	declarations: [
@@ -26,7 +27,10 @@ import {HeaderComponent} from './project/header/header.component';
 		AdminPageComponent,
 		AdminHeaderComponent,
 		AdminFooterComponent,
+		AdminContentComponent,
 		AdminSidebarComponent,
+		ProjectComponent,
+		FooterComponent
 	],
 	imports: [
 		BrowserModule,
