+++ frontend/src/app/layouts/admin/admin-page/admin-page.component.scss
@@ -2,6 +2,34 @@
 
 .page {
   display: flex;
+  background-color: #edf4f8;
+  min-height: calc(100% - 58px);
+  &-header {
+    background-color: #fff;
+    box-shadow: 2px 2px 2px rgba(0,0,0,0.1);
+    &__title {
+      padding: 0 15px;
+      font-size: 18px;
+      line-height: 62px;
+    }
+  }
 }
-=======
->>>>>>> admin-layout
+app-admin-sidebar {
+  min-width: 90px;
+  max-width: 90px;
+  background-color: white;
+  box-shadow: 1px 1px 1px rgba(0,0,0,0.1);
+}
+.sidebar-hidden {
+  display: none;
+}
+.inner-container {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+}
+app-admin-content {
+  height: 100%;
+  padding: 50px 15px;
+}
