+++ frontend/src/app/layouts/admin/admin-page/admin-page.component.ts
@@ -1,15 +1,17 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-admin-page',
-  templateUrl: './admin-page.component.html',
-  styleUrls: ['./admin-page.component.scss']
+	selector: 'app-admin-page',
+	templateUrl: './admin-page.component.html',
+	styleUrls: ['./admin-page.component.scss']
 })
 export class AdminPageComponent implements OnInit {
+	sidebarIsCollapsed = false;
+	constructor() { }
 
-  constructor() { }
-
-  ngOnInit() {
-  }
-
+	ngOnInit() {
+	}
+	toggleSidebar() {
+		this.sidebarIsCollapsed = !this.sidebarIsCollapsed;
+	}
 }
