+++ frontend/src/app/layouts/admin/admin-page/content/content.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+	selector: 'app-admin-content',
+	templateUrl: './content.component.html',
+	styleUrls: ['./content.component.scss']
+})
+export class AdminContentComponent implements OnInit {
+
+	constructor() { }
+
+	ngOnInit() {
+	}
+
+}
