+++ frontend/src/app/layouts/admin/admin-page/footer/footer.component.html
@@ -1,3 +1,3 @@
-<p>
-  footer works!
+<p class="footer-copyright">
+  &copy; Scripty 2018
 </p>
