+++ frontend/src/app/layouts/admin/admin-page/footer/footer.component.scss
@@ -1,0 +1,7 @@
+.footer-copyright {
+  padding: 0 30px;
+  line-height: 41px;
+  background-color: #212529;
+  font-size: 12px;
+  color: #fff;
+}
