+++ frontend/src/app/layouts/admin/admin-page/header/header.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import {AdminPageComponent} from '../admin-page.component';
 
 @Component({
 	selector: 'app-admin-header',
@@ -6,10 +7,12 @@ import { Component, OnInit } from '@angular/core';
 	styleUrls: ['./header.component.scss']
 })
 export class AdminHeaderComponent implements OnInit {
-
-	constructor() { }
-
+	sidebarIsCollapsed: boolean;
+	constructor(private AdminParent: AdminPageComponent) { }
 	ngOnInit() {
 	}
-
+	toggleSidebar () {
+		this.AdminParent.toggleSidebar();
+		this.sidebarIsCollapsed = this.AdminParent.sidebarIsCollapsed;
+	}
 }
