+++ frontend/src/app/layouts/admin/admin-page/sidebar/sidebar.component.html
@@ -1,32 +1,24 @@
 <div class="side-navbar">
   <div class="sidebar-header">
-    <div class="sidebar-avatar">
-      <img src="https://avatars0.githubusercontent.com/u/1?v=4" alt="">
-    </div>
-    <div class="title">
-      <h4 class="title-user">
-        UserName UserLastName
-      </h4>
-      <p class="title-position">Web Designer</p>
-    </div>
+		<img class="sidebar-avatar" src="https://avatars0.githubusercontent.com/u/1?v=4" alt="">
   </div>
   <span class="sidebar-category">Main</span>
   <ul class="sidebar-sections">
-    <li class="sidebar-item">
-      <a href="" class="sidebar-link">
-        <i class="fas fa-home"></i>
+    <li class="sidebar-item sidebar-item--active">
+      <a class="sidebar-link">
+        <i class="fas fa-home sidebar-icon"></i>
         Home
       </a>
     </li>
     <li class="sidebar-item">
-      <a href="" class="sidebar-link">
-        <i class="fas fa-table"></i>
+      <a class="sidebar-link">
+        <i class="fas fa-table sidebar-icon"></i>
         Tables
       </a>
     </li>
     <li class="sidebar-item">
-      <a href="" class="sidebar-link">
-        <<i class="far fa-chart-bar"></i>
+      <a class="sidebar-link">
+        <i class="far fa-chart-bar sidebar-icon"></i>
         Charts
       </a>
     </li>
