+++ frontend/src/app/layouts/admin/admin-page/sidebar/sidebar.component.scss
@@ -1,4 +1,38 @@
-.side-navbar {
-  min-width: 90px;
-  max-width: 90px;
+.sidebar {
+  &-avatar {
+    width: calc(100% - 30px);
+    padding: 30px 15px;
+    border-radius: 50%;
+  }
+  &-category {
+    display: block;
+    text-align: center;
+    text-transform: uppercase;
+    color: #ccc;
+    font-size: 12px;
+  }
+  &-sections {
+    margin-top: 15px;
+  }
+  &-item {
+    padding: 15px 2px;
+    &--active {
+      background-color: #0f84fd;
+      & > a, i {
+        color: #fff;
+      }
+    }
+  }
+  &-link {
+    display: block;
+    color: #aaa;
+    text-align: center;
+    font-size: 12px;
+  }
+  &-icon {
+    display: block;
+    margin-bottom: 10px;
+    color: #000;
+    font-size: 18px;
+  }
 }
