+++ frontend/src/app/layouts/admin/admin-page/sidebar/sidebar.component.ts
@@ -1,15 +1,13 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
 
 @Component({
-  selector: 'app-admin-sidebar',
-  templateUrl: './sidebar.component.html',
-  styleUrls: ['./sidebar.component.scss']
+	selector: 'app-admin-sidebar',
+	templateUrl: './sidebar.component.html',
+	styleUrls: ['./sidebar.component.scss']
 })
 export class AdminSidebarComponent implements OnInit {
 
-  constructor() { }
-
-  ngOnInit() {
-  }
-
+	constructor() { }
+	ngOnInit() {
+	}
 }
