--- frontend/src/app/layouts/home/header/header.component.html
@@ -1,15 +1,0 @@
-<div class="header">
-    <ul>
-        <li>
-            <img src="../../../../assets/img/icon.png" alt="Icon">
-        </li>
-        <li (click)="showIframe()"><i class="fas fa-caret-right"> Run</i></li>
-        <li><i class="fas fa-cloud-upload-alt"> Save</i></li>
-        <li><i class="fas fa-align-left"> Tidy</i></li>
-        <li><i class="fas fa-comment-alt"> Collaborate</i></li>
-    </ul>     
-    <ul>
-        <li><i class="fas fa-cogs"> Settings</i></li>
-        <li><b>Sign in</b></li>
-    </ul>   
-</div>
