--- frontend/src/app/layouts/home/header/header.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-	selector: 'app-header',
-	templateUrl: './header.component.html',
-	styleUrls: ['./header.component.scss']
-})
-export class HeaderComponent implements OnInit {
-
-	constructor() { }
-
-	ngOnInit() {
-	}
-
-}
