--- frontend/src/app/layouts/home/home.component.html
@@ -1,5 +1,0 @@
-<app-header></app-header>
-<main>
-  <app-sidebar></app-sidebar>
-  <app-main></app-main>
-</main>
