--- frontend/src/app/layouts/home/main/main.component.html
@@ -1,17 +1,0 @@
-<div class="mainContent">
-	<form>
-		<p>HTML <i class="fas fa-caret-down"></i></p>
-		<textarea></textarea>
-	</form>
-	<form>
-		<p>CSS <i class="fas fa-caret-down"></i></p>
-		<textarea></textarea>
-	</form>
-	<form>
-		<p>Javascript <i class="fas fa-caret-down"></i></p>
-		<textarea></textarea>
-	</form>
-	<form>
-		<p>Result</p>
-	</form>
-</div>
