--- frontend/src/app/layouts/home/main/main.component.scss
@@ -1,38 +1,0 @@
-@import "../../../variables";
-
-.mainContent{
-    width: 87vw;
-    height: 93vh;
-    background-color: $primary-color;
-    border: 2px solid $border-color; 
-    display: flex;
-    flex-wrap: wrap;
-
-    form{
-        iframe{
-            display: none;
-        }
-
-        border: 2px solid $border-color; 
-        display: flex;
-        flex-direction: column;
-        padding: 5px;
-        width: 48.5%;
-
-        i,p{
-            color: white;
-        }
-        p{
-            font-weight: 100;
-        }
-        textarea{
-            border: 0;
-            height: 100%;
-            background-color: $primary-color;
-            color: white;
-            font-size: 1em;
-            resize: none;
-            caret-color: yellow;
-        }
-    }
-}
