--- frontend/src/app/layouts/home/main/main.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-	selector: 'app-main',
-	templateUrl: './main.component.html',
-	styleUrls: ['./main.component.scss']
-})
-export class MainComponent {}
