--- frontend/src/app/layouts/home/sidebar/sidebar.component.html
@@ -1,1 +1,0 @@
-<div class="sidebar"></div>
