--- frontend/src/app/layouts/home/sidebar/sidebar.component.scss
@@ -1,8 +1,0 @@
-@import "../../../variables";
-
-.sidebar{
-    width: 13vw;
-    height: 92.6vh;
-    background-color: $primary-color;
-    border: 2px solid $border-color; 
-}
