+++ frontend/src/app/layouts/project/footer/footer.component.scss
