+++ frontend/src/app/layouts/project/footer/footer.component.spec.ts
