+++ frontend/src/app/layouts/project/header/header.component.html
@@ -1,6 +1,6 @@
 <header>
     <a href="/" class="logo">
-        <img src="../../assets/img/icon.png" alt="logo">
+        <img src="../../../../assets/img/icon.png" alt="logo">
     </a>
     <nav [ngClass]=" isOpen? 'open':''">
         <ul class="action-cont">
