+++ frontend/src/app/layouts/project/main/main.component.scss
