+++ frontend/src/app/layouts/project/project.component.scss
