+++ frontend/src/app/layouts/project/sidebar/sidebar.component.ts
